test: don't import the OpenTelemetry SDK in the realtime helper tests
The SDK only ships with the optional `logfire` extra, so importing it at module
level broke collection in the `pydantic-ai-slim` and `pydantic-evals` CI jobs.
Move the two `inject_trace_context` tests to `test_instrumentation.py`, which
already gates on it with `pytest.importorskip`.
